iot-sensor-system/requirements.txt
# Core
flask==3.0.0
python-dotenv==1.0.0

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Bases de datos
pymongo==4.6.0
psycopg2-binary==2.9.9
redis==5.0.1

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# Visualización y API
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Autenticación
PyJWT==2.8.0
bcrypt==4.1.1

# Utilidades
python-dateutil==2.8.2
requests==2.31.0